requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
razorpay==1.4.2
psycopg2-binary==2.9.6